+++ nodes/SeaTable/v2/actions/row/index.ts
@@ -1,6 +1,7 @@
 import type { INodeProperties } from 'n8n-workflow';
 
 import * as create from './create.operation';
+import * as upsert from './upsert.operation';
 import * as get from './get.operation';
 import * as list from './list.operation';
 import * as lock from './lock.operation';
@@ -9,7 +10,7 @@ import * as search from './search.operation';
 import * as unlock from './unlock.operation';
 import * as update from './update.operation';
 
-export { create, get, search, update, remove, lock, unlock, list };
+export { create, upsert, get, search, update, remove, lock, unlock, list };
 
 export const descriptions: INodeProperties[] = [
 	{
@@ -28,6 +29,12 @@ export const descriptions: INodeProperties[] = [
 				value: 'create',
 				description: 'Create a new row',
 				action: 'Create a row',
+			},
+						{
+				name: 'Create or Update',
+				value: 'upsert',
+				description: 'Create a new row, or update the current one if it already exists (upsert)',
+				action: 'Create or update a row',
 			},
 			{
 				name: 'Delete',
@@ -75,6 +82,7 @@ export const descriptions: INodeProperties[] = [
 		default: 'create',
 	},
 	...create.description,
+	...upsert.description,
 	...get.description,
 	...list.description,
 	...search.description,